Add ieasemusic

+++ build-deb.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 
-## Build environments
 
 ## Step 1: Enter directory
 HERE="$(dirname "$(readlink -f "${0}")")"
@@ -31,6 +30,17 @@ parse_yaml() {
 eval $(parse_yaml recipe.yml "_")
 
 
+## Step 3: Check if we need to skip this build
+if [ -n "${_only_arches[@]}" ]; then
+    for BUILD_ARCH in "${_only_arches[@]}" ; do
+        if [ "$BUILD_ARCH" = "$DEBIAN_ARCH" ]; then
+            echo "\e[32m *** Skip build for $_name *** \e[0m"
+            exit 0
+        fi
+    done
+fi
+
+
 ## Step 3: Get latest version and source url
 get_latest_version_github() {
     export PYTHONIOENCODING=utf8
@@ -40,9 +50,13 @@ get_latest_version_github() {
 
 if [ "$_source_host" = "github" ]; then
     LATEST_VERSION=$(get_latest_version_github "$_source_repo")
-    SOURCE_URL="https://github.com/$_source_repo/archive/$LATEST_VERSION.tar.gz"
     LATEST_VERSION="${LATEST_VERSION#v}"
-    SOURCE_DIR="${_name}-${LATEST_VERSION}"
+    if [ "$_source_method" = "build" ]; then
+        SOURCE_URL="https://github.com/$_source_repo/archive/v$LATEST_VERSION.tar.gz"
+        SOURCE_DIR="${_name}-${LATEST_VERSION}"
+    else
+        eval "PACKAGE_URL=${_source_package_url}"
+    fi
 else
     echo "Error: unsupported host: $_source_host" > /dev/stderr
     exit 1
@@ -50,38 +64,44 @@ fi
 
 
 ## Step 4: Check if the package of latest version exists
-PACKAGE_URL="https://dl.bintray.com/coslyk/debianzh/pool/main/${_name:0:1}/${_name}/${_name}_${LATEST_VERSION}-1~${DEBIAN_RELEASE}_${DEBIAN_ARCH}.deb"
-if curl --output /dev/null --silent --head --fail "$PACKAGE_URL"; then
+REMOTE_URL="https://dl.bintray.com/coslyk/debianzh/pool/main/${_name:0:1}/${_name}/${_name}_${LATEST_VERSION}-1~${DEBIAN_RELEASE}_${DEBIAN_ARCH}.deb"
+if curl --output /dev/null --silent --head --fail "$REMOTE_URL"; then
     exit 0
 else
     echo -e "\e[32m *** Detect update for $_name: $LATEST_VERSION *** \e[0m"
 fi
 
 
-## Step 5: Download source code
-echo "Downloading source code from $SOURCE_URL"
-curl -L -o source.tar.gz "$SOURCE_URL"
-tar xzf source.tar.gz
-
-
-## Step 6: Create debian directory
+## Step 5: Download source code or package
 if [ "$_source_method" = "build" ]; then
+    echo "Downloading source code from $SOURCE_URL"
+    curl -L -o source.tar.gz "$SOURCE_URL"
+    tar xzf source.tar.gz
+else
+    echo "Downloading $PACKAGE_URL"
+    curl -L -o package.deb "$PACKAGE_URL"
+fi
+
+
+## Step 6: Build package if needed
+if [ "$_source_method" = "build" ]; then
+
+    # Copy debian folder
     [ -d $SOURCE_DIR/debian ] && rm -rf $SOURCE_DIR/debian
     cp -r debian-template $SOURCE_DIR/debian
     find $SOURCE_DIR/debian -type f -exec sed -i -e "s|##VERSION|$LATEST_VERSION|g" {} \;
     find $SOURCE_DIR/debian -type f -exec sed -i -e "s|##RELEASE|$DEBIAN_RELEASE|g" {} \;
+    
+    # Build package
+    printf "Building "
+    $HERE/travis/build -i docker-deb-builder:buster -o . $SOURCE_DIR | while read LINE; do
+        printf "."
+    done
+    printf "\n"
+    rm -f *-dbgsym_*.deb
+    printf "Built: "
+    ls *.deb
 fi
-
-
-## Step 7: Build package
-printf "Building "
-$HERE/travis/build -i docker-deb-builder:buster -o . $SOURCE_DIR | while read LINE; do
-    printf "."
-done
-printf "\n"
-rm -f *-dbgsym_*.deb
-printf "Built: "
-ls *.deb
 
 # Step 9: Upload
 echo "Uploading ..."
